Small website built on HTML in 4 hours
This is a small website I built using HTML and some CSS code so as to better understand how these two things work in todays web. I also included a small CTF in the code, if you want to have a go at it you'll have to check out the contact page. Some of the hyperlinks on the pages may need repointing for your file directory if you download these files.